+++ procesador/procesador.ts
@@ -6,13 +6,12 @@ import type {
   LLavesMultiples,
   Listas,
   LllavesSingulares,
-  Proyecto,
-  Lugar,
-  elementoGeoJson
+  Proyecto
 } from '../src/tipos.ts';
 import { getXlsxStream } from 'xlstream';
 import slugificar from 'slug';
-import { guardarJSON, guardarGEOJSON, ordenarListaObjetos } from './ayudas.js';
+import { guardarJSON, ordenarListaObjetos } from './ayudas.js';
+import procesarLugares from './lugares.js';
 
 const datosEmpiezanEnFila = 2;
 const camposSingulares: Campos = [
@@ -34,7 +33,6 @@ const campos = [...camposSingulares, ...camposMultiples];
 const proyectos: Proyecto[] = [];
 
 const listas: Listas = {
-  id: [],
   regiones: [],
   años: [],
   tipos: [],
@@ -48,160 +46,143 @@ const listas: Listas = {
   decadas: []
 };
 
-const lugares: Lugar[] = [];
-const features: Object[] = [];
-
-let geojson = { type: 'FeatureCollection', features: features };
-
-procesar();
+const archivo = './procesador/datos/Listado de proyectos - 60 años dpto antropología0802.xlsx';
 
 async function procesar() {
-  const flujo = await getXlsxStream({
-    filePath: './procesador/Listado de proyectos - 60 años dpto antropología0802.xlsx',
-    sheet: 'Proyectos',
-    withHeader: false,
-    ignoreEmpty: true
-  });
+  await procesarProyectos();
+  console.log('Proyectos procesados');
+  await procesarLugares(archivo, listas);
+  console.log('fin de lugares');
 
-  const flujoLugares = await getXlsxStream({
-    filePath: './procesador/Listado de proyectos - 60 años dpto antropología0802.xlsx',
-    sheet: 'lugares',
+  console.log('fin');
+}
+
+procesar().catch(console.error);
+
+async function procesarProyectos(): Promise<void> {
+  const flujo = await getXlsxStream({
+    filePath: archivo,
+    sheet: 'Proyectos',
     withHeader: false,
     ignoreEmpty: true
   });
 
   let numeroFila = 1;
 
-  flujoLugares.on('data', (fila) => {
-    if (numeroFila > 2) {
-      procesarLugar(fila.formatted.arr);
-    } else {
-    }
+  return new Promise((resolver, rechazar) => {
+    flujo.on('data', (fila) => {
+      if (numeroFila > datosEmpiezanEnFila) {
+        procesarFila(fila.formatted.arr);
+      }
 
-    numeroFila++;
-  });
+      numeroFila++;
+    });
 
-  flujo.on('data', (fila) => {
-    if (numeroFila > datosEmpiezanEnFila) {
-      procesarFila(fila.formatted.arr);
-    } else {
-    }
+    flujo.on('close', () => {
+      for (const lista in listas) {
+        ordenarListaObjetos(listas[lista as keyof Listas], 'slug', true);
+      }
 
-    numeroFila++;
-  });
+      proyectos.forEach((proyecto) => {
+        const id = proyecto.id;
 
-  flujo.on('close', () => {
-    for (const lista in listas) {
-      ordenarListaObjetos(listas[lista as keyof Listas], 'slug', true);
-    }
+        campos.forEach((campoRelacion) => {
+          const datosRelacion = proyecto[campoRelacion.llave];
 
-    proyectos.forEach((proyecto) => {
-      const id = proyecto.id;
+          campos.forEach((campo) => {
+            // Agregar datos de cada campo en todos los otros, excepto en si mismo.
+            if (campoRelacion.llave !== campo.llave && datosRelacion) {
+              const llaveALlenar = campo.llave;
+              const llaveDondeLllenar = campoRelacion.llave;
+              const datosProyecto = proyecto[llaveALlenar];
 
-      campos.forEach((campoRelacion) => {
-        const datosRelacion = proyecto[campoRelacion.llave];
+              // Si el proyecto tiene datos en este campo
+              if (datosProyecto) {
+                // Sacar los slugs del campo
+                const slugsCampoProyecto = Array.isArray(datosProyecto)
+                  ? (datosProyecto as DefinicionSimple[]).map(({ slug }) => slug)
+                  : [(datosProyecto as DefinicionSimple).slug];
 
-        campos.forEach((campo) => {
-          // Agregar datos de cada campo en todos los otros, excepto en si mismo.
-          if (campoRelacion.llave !== campo.llave && datosRelacion) {
-            const llaveALlenar = campo.llave;
-            const llaveDondeLllenar = campoRelacion.llave;
-            const datosProyecto = proyecto[llaveALlenar];
+                slugsCampoProyecto.forEach((slug) => {
+                  const i = listas[llaveALlenar].findIndex((obj) => obj.slug === slug);
+                  const elementosDondeConectar = Array.isArray(datosRelacion)
+                    ? (datosRelacion as DefinicionSimple[]).map(({ slug }) => slug)
+                    : [datosRelacion.slug];
 
-            // Si el proyecto tiene datos en este campo
-            if (datosProyecto) {
-              // Sacar los slugs del campo
-              const slugsCampoProyecto = Array.isArray(datosProyecto)
-                ? (datosProyecto as DefinicionSimple[]).map(({ slug }) => slug)
-                : [(datosProyecto as DefinicionSimple).slug];
+                  elementosDondeConectar.forEach((elementoConector) => {
+                    const elementoALlenar = listas[llaveDondeLllenar].find((obj) => obj.slug === elementoConector);
 
-              slugsCampoProyecto.forEach((slug) => {
-                const i = listas[llaveALlenar].findIndex((obj) => obj.slug === slug);
-                const elementosDondeConectar = Array.isArray(datosRelacion)
-                  ? (datosRelacion as DefinicionSimple[]).map(({ slug }) => slug)
-                  : [datosRelacion.slug];
+                    if (elementoALlenar) {
+                      const existe = elementoALlenar.relaciones.find((obj) => obj.slug === slug);
 
-                elementosDondeConectar.forEach((elementoConector) => {
-                  const elementoALlenar = listas[llaveDondeLllenar].find((obj) => obj.slug === elementoConector);
+                      if (!elementoALlenar.proyectos.includes(id)) {
+                        elementoALlenar.proyectos.push(id);
+                      }
 
-                  if (elementoALlenar) {
-                    const existe = elementoALlenar.relaciones.find((obj) => obj.slug === slug);
-
-                    if (!elementoALlenar.proyectos.includes(id)) {
-                      elementoALlenar.proyectos.push(id);
+                      if (!existe) {
+                        elementoALlenar.relaciones.push({
+                          conteo: 1,
+                          indice: i,
+                          tipo: llaveALlenar,
+                          slug
+                        });
+                      } else {
+                        existe.conteo++;
+                      }
                     }
+                  });
+                });
+              }
+            }
+          });
+        });
+
+        // Años
+        if (proyecto.años) {
+          proyecto.años.años.forEach((año) => {
+            const añoEnSlug = `${año}`;
+            campos.forEach((campo) => {
+              const datosCampo = proyecto[campo.llave];
+              const llaveALlenar = campo.llave;
+              if (datosCampo) {
+                const slugs = Array.isArray(datosCampo) ? datosCampo.map(({ slug }) => slug) : [datosCampo.slug];
+
+                slugs.forEach((slug) => {
+                  const objALlenar = listas[llaveALlenar].find((obj) => obj.slug === slug);
+
+                  if (objALlenar) {
+                    const i = listas.años.findIndex((año) => año.slug === añoEnSlug);
+                    const existe = objALlenar.relaciones.find((obj) => obj.slug === añoEnSlug && obj.tipo === 'años');
 
                     if (!existe) {
-                      elementoALlenar.relaciones.push({
-                        conteo: 1,
-                        indice: i,
-                        tipo: llaveALlenar,
-                        slug
-                      });
+                      objALlenar.relaciones.push({ conteo: 1, indice: i, tipo: 'años', slug: añoEnSlug });
                     } else {
                       existe.conteo++;
                     }
                   }
+
+                  // Llenar este en años
+                  const objAño = listas.años.find((obj) => obj.slug === añoEnSlug);
+
+                  if (objAño) {
+                    const i = listas[llaveALlenar].findIndex((obj) => obj.slug === slug);
+                    const existe = objAño.relaciones.find((obj) => obj.slug === slug);
+
+                    if (!existe) {
+                      objAño.relaciones.push({ conteo: 1, indice: i, tipo: llaveALlenar, slug });
+                    }
+                  }
                 });
-              });
-            }
-          }
-        });
+              }
+            });
+          });
+        }
       });
 
-      // Años
-      if (proyecto.años) {
-        proyecto.años.años.forEach((año) => {
-          const añoEnSlug = `${año}`;
-          campos.forEach((campo) => {
-            const datosCampo = proyecto[campo.llave];
-            const llaveALlenar = campo.llave;
-            if (datosCampo) {
-              const slugs = Array.isArray(datosCampo) ? datosCampo.map(({ slug }) => slug) : [datosCampo.slug];
-
-              slugs.forEach((slug) => {
-                const objALlenar = listas[llaveALlenar].find((obj) => obj.slug === slug);
-
-                if (objALlenar) {
-                  const i = listas.años.findIndex((año) => año.slug === añoEnSlug);
-                  const existe = objALlenar.relaciones.find((obj) => obj.slug === añoEnSlug && obj.tipo === 'años');
-
-                  if (!existe) {
-                    objALlenar.relaciones.push({ conteo: 1, indice: i, tipo: 'años', slug: añoEnSlug });
-                  } else {
-                    existe.conteo++;
-                  }
-                }
-
-                // Llenar este en años
-                const objAño = listas.años.find((obj) => obj.slug === añoEnSlug);
-
-                if (objAño) {
-                  const i = listas[llaveALlenar].findIndex((obj) => obj.slug === slug);
-                  const existe = objAño.relaciones.find((obj) => obj.slug === slug);
-
-                  if (!existe) {
-                    objAño.relaciones.push({ conteo: 1, indice: i, tipo: llaveALlenar, slug });
-                  }
-                }
-              });
-            }
-          });
-        });
-      }
+      guardarJSON(proyectos, 'proyectos');
+      guardarJSON(listas, 'listas');
+      resolver();
     });
-
-    guardarJSON(proyectos, 'proyectos');
-    guardarJSON(listas, 'listas');
-
-    console.log('fin');
-  });
-
-  flujoLugares.on('close', () => {
-    procesarDatosMapa();
-    guardarGEOJSON(geojson, 'datosMapa');
-
-    console.log('fin de lugares');
   });
 }
 
@@ -225,27 +206,6 @@ function procesarFila(fila: string[]) {
   });
 
   proyectos.push(respuesta);
-}
-
-function procesarLugar(fila: string[]) {
-  const nombreLugar = fila[0].trim();
-  const slug = slugificar(nombreLugar);
-  const longitud = fila[8];
-  const latitud = fila[7];
-  const lugar = listas.municipios.filter((elemento) => elemento.slug === slug);
-  const conteo = lugar[0] ? lugar[0].conteo : 0;
-
-  const respuesta: Lugar = {
-    nombre: nombreLugar,
-    slug: slug,
-    lon: +longitud,
-    lat: +latitud,
-    conteo: conteo
-  };
-
-  if (respuesta.lon && respuesta.lat) {
-    lugares.push(respuesta);
-  }
 }
 
 function validarValorMultiple(valor: string, lista: ElementoLista[]) {
@@ -311,23 +271,4 @@ function validarAño(valorAño: string) {
   }
 
   return añoProcesado;
-}
-
-// Función para crear geojson con lugares y cantidad de proyectos por lugar
-function procesarDatosMapa() {
-  let elemento: elementoGeoJson;
-  for (let lugar in lugares) {
-    elemento = {
-      type: 'Feature',
-      properties: {
-        slug: lugares[lugar].slug,
-        conteo: lugares[lugar].conteo
-      },
-      geometry: { type: 'Point', coordinates: [lugares[lugar].lon, lugares[lugar].lat] }
-    };
-
-    if (elemento.properties.conteo > 0) {
-      features.push(elemento);
-    }
-  }
 }
